% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aceEndicoFunctions.R
\name{aeInvoiceCodeFix}
\alias{aeInvoiceCodeFix}
\title{Corrects invoice codes, since Ace Endico can't seem to do it themselves.
Normally most AE invoice codes should be formatted 'X00000-00', but for some
reason some are formatted without the hyphen or with lower case letters. This
fixes that in cases where the code is not correct.}
\usage{
aeInvoiceCodeFix(df, var)
}
\arguments{
\item{df}{This is an import of the Ace Endico reports containing an invoice
value}

\item{var}{The invoice variable that needs to be cleaned}
}
\value{
A dataframe containing a corrected invoice variable
}
\description{
This function cleans Invoice codes on Aced Endico Reports, since they can't
do it themselves.
}
