% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.R
\name{bps_add_date}
\alias{bps_add_date}
\title{Create date column in dataset}
\usage{
bps_add_date(data)
}
\arguments{
\item{data}{A tibble with the \code{bpsr_tbl} subclass.}
}
\value{
A tibble that has a \code{date} column.
}
\description{
Adds a \code{date} variable to a BPS dataset, based on years and period IDs or
numeric representations of months. The values are the floor date of the
given period. The format is year-month-date (ymd).
}
\examples{
\dontshow{if (httr2::secret_has_key("BPSR_KEY")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
data <- bps_get_trade_hs_chapter("01", "export", year = 2021)
bps_add_date(data)
\dontshow{\}) # examplesIf}
}
\seealso{
\itemize{
\item \code{\link[=bps_get_dataset]{bps_get_dataset()}} and \code{\link[=bps_get_datasets]{bps_get_datasets()}} to request datasets.
\item \code{\link[=bps_get_trade_hs_chapter]{bps_get_trade_hs_chapter()}} and \code{\link[=bps_get_trade_hs_full]{bps_get_trade_hs_full()}} to request
trade datasets.
}
}
