% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-spreadsheet.R
\name{bps_dataset_spreadsheet}
\alias{bps_dataset_spreadsheet}
\alias{bps_download_spreadsheet}
\title{Look up and request spreadsheet dataset}
\usage{
bps_dataset_spreadsheet(
  keyword = NULL,
  month = NULL,
  year = NULL,
  domain_id = "0000",
  page = NULL,
  lang = c("ind", "eng")
)

bps_download_spreadsheet(
  dataset_id,
  file_path,
  domain_id = "0000",
  lang = c("ind", "eng")
)
}
\arguments{
\item{keyword}{The keyword to filter the search result.}

\item{month}{The month in which a spreadsheet is last updated.}

\item{year}{The year in which a spreadsheet is last updated.}

\item{domain_id}{The domain ID of a BPS bureau. Defaults to "0000", which
is the domain ID of the central bureau. Use \code{\link[=bps_domain]{bps_domain()}} to see the
list of domain IDs.}

\item{page}{The page number of the result. Defaults to page one. Set to \code{Inf}
to request all pages.}

\item{lang}{The result's language. Must be either "ind" for Indonesian or
"eng" for English. Defaults to Indonesian.}

\item{dataset_id}{The dataset ID. Use \code{\link[=bps_dataset_spreadsheet]{bps_dataset_spreadsheet()}} to see
the list of dataset IDs.}

\item{file_path}{The path to save the spreadsheet. Must have a ".xls"
extension.}
}
\value{
\itemize{
\item \code{bps_dataset_spreadsheet()} returns a tibble containing dataset IDs and
titles, as well as other related details.
\item \code{bps_download_spreadsheet()} invisibly returns a string of the file path.
}
}
\description{
Search and download datasets that BPS distributes in spreadsheets. The API
calls them static tables.
\itemize{
\item \code{bps_dataset_spreadsheet()} requests the spreadsheet dataset table
\item \code{bps_download_spreadsheet()} downloads a spreadsheet
}
}
\section{File format}{
The spreadsheet dataset comes in a Microsoft Excel file format with the
".xls" extension.
}

\examples{
\dontshow{if (httr2::secret_has_key("BPSR_KEY")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Search spreadsheet datasets on unemployment
spreadsheet <- bps_dataset_spreadsheet(
  keyword = "unemployment",
  lang = "eng"
)

\dontrun{
bps_download_spreadsheet(spreadsheet$dataset_id[[1]], "spreadsheet.xls")
}
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=bps_download]{bps_download()}} to download resources by URL.
}
