% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain.R
\name{bps_domain}
\alias{bps_domain}
\title{Look up domain of BPS bureau}
\usage{
bps_domain(level = c("all", "province", "city"))
}
\arguments{
\item{level}{The level of BPS bureaus. Must be either "all", "province" or
"city". Defaults to "all".}
}
\value{
A tibble containing domain IDs, names and URLs of every BPS bureau
at the given \code{level}.
}
\description{
Request the table of domains of BPS bureaus at the city or regency level,
province level or both levels.
}
\examples{
\dontshow{if (httr2::secret_has_key("BPSR_KEY")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# See the list of domains of BPS bureaus at the province level
bps_domain(level = "province")
\dontshow{\}) # examplesIf}
}
