% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{bps_download}
\alias{bps_download}
\title{Save resource from BPS API}
\usage{
bps_download(url, file_path)
}
\arguments{
\item{url}{A BPS API download URL. Supply a vector of URLs to download
multiple resources.}

\item{file_path}{The path to save the resource. Must have a ".xls", ".pdf" or
".png" extension. Supply a vector of the same length as \code{url} to
save multiple resources.}
}
\value{
A character vector of \code{file_path}.
}
\description{
Download resources from BPS API, such as spreadsheet datasets, publications,
press releases or infographics.
}
\examples{
\dontshow{if (httr2::secret_has_key("BPSR_KEY")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Download an infographic
infographics <- bps_search("infographic")

\dontrun{
bps_download(infographics$file_url[[1]], "infographics.png")
}
\dontshow{\}) # examplesIf}
}
\seealso{
\itemize{
\item \code{\link[=bps_view]{bps_view()}} to get information such as the download URL of a resource.
\item \code{\link[=bps_search]{bps_search()}} to find resources and their download URLs.
\item \code{\link[=bps_dataset_spreadsheet]{bps_dataset_spreadsheet()}} to look up spreadsheet datasets and their
download URLs.
}
}
