% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade.R
\name{bps_get_trade_hs_chapter}
\alias{bps_get_trade_hs_chapter}
\alias{bps_get_trade_hs_full}
\alias{bps_hs_code}
\alias{bps_guess_latest_trade_year}
\title{Request trade data}
\usage{
bps_get_trade_hs_chapter(
  hs_chapter,
  flow,
  year = bps_guess_latest_trade_year(),
  freq = c("monthly", "annual")
)

bps_get_trade_hs_full(
  hs_code,
  flow,
  year = bps_guess_latest_trade_year(),
  freq = c("monthly", "annual")
)

bps_hs_code()

bps_guess_latest_trade_year()
}
\arguments{
\item{hs_chapter, hs_code}{The HS chapter or code. To request by multiple
chapters or codes, supply a vector of them.}

\item{flow}{The trade flow or direction. Must be either "export" or "import".}

\item{year}{The year of observation. Defaults to
\code{bps_guess_latest_trade_year()}, which will try to guess the year in which
the latest observations are available.

Set to \code{Inf} to request data across all years.

To request by multiple years, supply a vector of the years.
See the Scope section below for more details.}

\item{freq}{The frequency of observation. Must be either "monthly" or
"annual". Defaults to "monthly".}
}
\value{
\itemize{
\item \code{bps_hs_code()} returns a tibble containing the HS chapters and codes, as
well as their description.
\item \code{bps_get_trade_hs_chapter()} and \code{bps_get_trade_hs_full()} return a tibble
with the \code{bpsr_tbl} subclass, which has a \code{metadata} attribtue. Use
\code{\link[=bps_metadata]{bps_metadata()}} to read it.
\item \code{bps_guess_latest_trade_year()} returns an integer.
}
}
\description{
Get data on export from and import to Indonesia.
\itemize{
\item \code{bps_hs_code()} retrieves the Harmonized System (HS) code table. This table
contains the HS chapters and codes, which are needed to request trade
data
\item \code{bps_get_trade_hs_chapter()} requests export or import data by the
HS chapter or the first two digits of the HS code
\item \code{bps_get_trade_hs_full()} requests export or import data by the HS code
\item \code{bps_guess_latest_trade_year()} gives the estimated year in which the most
recent data are available
}
}
\section{Scope}{
The API provides export and import data only as far back as 2014.
}

\examples{
\dontshow{if (httr2::secret_has_key("BPSR_KEY")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Request by the HS chapter
bps_get_trade_hs_chapter("03", "export", year = 2018)
\dontshow{\}) # examplesIf}
}
