% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{bps_list}
\alias{bps_list}
\title{Perform request to list endpoint}
\usage{
bps_list(model, ..., domain_id = "0000", page = NULL, lang = c("ind", "eng"))
}
\arguments{
\item{model}{The result model, e.g. "publication" to request a list of
publications. See the Model section below for more details.}

\item{...}{Name-value pairs that provide query parameters. See
\code{\link[httr2:req_url]{httr2::req_url_query()}} for more details.}

\item{domain_id}{The domain ID of a BPS bureau. Defaults to "0000", which
is the domain ID of the central bureau. Use \code{\link[=bps_domain]{bps_domain()}} to see the
list of domain IDs.}

\item{page}{The page number of the result. Defaults to page one. Set to \code{Inf}
to request all pages.}

\item{lang}{The result's language. Must be either "ind" for Indonesian or
"eng" for English. Defaults to Indonesian.}
}
\value{
A list with the \code{bpsr_table} class.
}
\description{
The low-level function to send one or more requests, depending on \code{page},
to the list endpoint.
}
\section{Model}{
\code{model} must be one of the following:
\itemize{
\item "subject";
\item "subcat";
\item "var";
\item "vervar";
\item "turvar";
\item "turth";
\item "th";
\item "unit";
\item "news";
\item "newscategory";
\item "pressrelease";
\item "publication";
\item "statictable";
\item "subcatcsa";
\item "subjectcsa";
\item "tablestatistic";
\item "indicators";
\item "infographic";
\item "glosarium";
\item "sdgs";
\item "kbli2009";
\item "kbli2015";
\item "kbli2017";
\item "kbli2020"; or
\item "kbki2015".
}

The API actually also accepts "data" for \code{model}, but bpsr puts it in a
separate family of functions due to differences in the response structure.
See \code{\link[=bps_get_dataset]{bps_get_dataset()}} and \code{\link[=bps_get_datasets]{bps_get_datasets()}} for more
details.
}

\examples{
\dontshow{if (httr2::secret_has_key("BPSR_KEY")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Get a list of some publications
bps_list("publication", page = c(1, 2, 3), lang = "eng")
\dontshow{\}) # examplesIf}
}
