% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{bps_metadata}
\alias{bps_metadata}
\title{Extract dataset metadata}
\usage{
bps_metadata(data)
}
\arguments{
\item{data}{A tibble with the \code{bpsr_tbl} subclass.}
}
\value{
A list with the \code{bpsr_metadata} class.
}
\description{
Read the metadata of a BPS dataset.
}
\examples{
\dontshow{if (httr2::secret_has_key("BPSR_KEY")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Read the metadata of personal care and other services inflation dataset
inflation <- bps_get_dataset("1927", lang = "eng")
bps_metadata(inflation)

# Read the metadata of a trade dataset
trade <- bps_get_trade_hs_chapter("01", "export")
bps_metadata(trade)
\dontshow{\}) # examplesIf}
}
\seealso{
\itemize{
\item \code{\link[=bps_get_dataset]{bps_get_dataset()}} and \code{\link[=bps_get_datasets]{bps_get_datasets()}} to request datasets.
\item \code{\link[=bps_get_trade_hs_chapter]{bps_get_trade_hs_chapter()}} and \code{\link[=bps_get_trade_hs_full]{bps_get_trade_hs_full()}} to request
trade datasets.
}
}
