% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{bps_request}
\alias{bps_request}
\alias{bps_request_multiple}
\alias{bps_request_paginated}
\title{Perform single, multiple or paginated request to BPS API}
\usage{
bps_request(url_path, ...)

bps_request_multiple(url_path, query_params)

bps_request_paginated(url_path, ..., page = 1)
}
\arguments{
\item{url_path}{The path component of the URL. Must be either "domain",
"list", "view", "dataexim" or "interoperabilitas".}

\item{...}{Name-value pairs that provide query parameters. See
\code{\link[httr2:req_url]{httr2::req_url_query()}} for more details.}

\item{query_params}{A list of name-value pairs that provide query parameters.}

\item{page}{The page number of the result. Defaults to page one. Set to \code{Inf}
to request all pages.}
}
\value{
\itemize{
\item \code{bps_request()} returns a list with the \code{bpsr_response} class.
\item \code{bps_request_multiple()} and \code{bps_request_paginated()} return a list with
the \code{bpsr_multiple_responses} class.
}
}
\description{
Perform both a single request and multiple requests to BPS API.
\itemize{
\item \code{bps_request()} sends a single request
\item \code{bps_request_multiple()} sends multiple requests with multiple query
parameters to one endpoint at a time
\item \code{bps_request_paginated()} sends multiple requests to multiple pages to
the list endpoint
}
}
\examples{
\dontshow{if (httr2::secret_has_key("BPSR_KEY")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Request a list of domains of every BPS bureau
bps_request("domain", type = "all")

# Request a list of dataset subjects on each domain on different pages
bps_request_multiple(
  "list",
  list(
    model = "subject",
    domain = c("0000", "1100", "1500"),
    page = c(1, 2, 1),
    lang = "eng"
  )
)

# Request the complete list of dataset subjects on the central bureau's
# domain
bps_request_paginated(
  "list",
  model = "subject",
  domain = "0000",
  lang = "eng",
  page = Inf
)
\dontshow{\}) # examplesIf}
}
