% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdg.R
\name{bps_sdg}
\alias{bps_sdg}
\title{Look up SDG dataset}
\usage{
bps_sdg(goal = NULL, page = NULL, lang = c("ind", "eng"))
}
\arguments{
\item{goal}{The goal number.}

\item{page}{The page number of the result. Defaults to page one. Set to \code{Inf}
to request all pages.}

\item{lang}{The result's language. Must be either "ind" for Indonesian or
"eng" for English. Defaults to Indonesian.}
}
\value{
A tibble containing dataset IDs and titles, among other things.
}
\description{
Request the table of datasets related to the Sustainable Development Goals
(SDG). This table contains dataset IDs, which are needed to request datasets
using \code{\link[=bps_get_dataset]{bps_get_dataset()}} or \code{\link[=bps_get_datasets]{bps_get_datasets()}}.
}
\examples{
\dontshow{if (httr2::secret_has_key("BPSR_KEY")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Get a list of some datasets related to goal number one
bps_sdg(goal = 1, lang = "eng")
\dontshow{\}) # examplesIf}
}
