% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{bps_search}
\alias{bps_search}
\title{Find resource on list endpoint}
\usage{
bps_search(
  model,
  ...,
  keyword = NULL,
  domain_id = "0000",
  page = NULL,
  lang = c("ind", "eng")
)
}
\arguments{
\item{model}{The result model, e.g. "publication" to request a list of
publications. See the Model section below for more details.}

\item{...}{Name-value pairs that provide query parameters. See
\code{\link[httr2:req_url]{httr2::req_url_query()}} for more details.}

\item{keyword}{The keyword to filter the search result.}

\item{domain_id}{The domain ID of a BPS bureau. Defaults to "0000", which
is the domain ID of the central bureau. Use \code{\link[=bps_domain]{bps_domain()}} to see the
list of domain IDs.}

\item{page}{The page number of the result. Defaults to page one. Set to \code{Inf}
to request all pages.}

\item{lang}{The result's language. Must be either "ind" for Indonesian or
"eng" for English. Defaults to Indonesian.}
}
\value{
A tibble.
}
\description{
Search resources on the list endpoint, such as press releases, news and
publications.
}
\section{Model}{
\code{model} must be one of the following:
\itemize{
\item "statictable";
\item "pressrelease";
\item "publication";
\item "infographic"; or
\item "news".
}
}

\examples{
\dontshow{if (httr2::secret_has_key("BPSR_KEY")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
bps_search("news", keyword = "inflation", lang = "eng")
\dontshow{\}) # examplesIf}
}
