% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key.R
\name{bps_set_key}
\alias{bps_set_key}
\title{Set BPS API key}
\usage{
bps_set_key(key = NULL)
}
\arguments{
\item{key}{The API key. Defaults to \code{NULL}, which prompts the user to add it
using a dialog box.}
}
\description{
Provide the API key to bpsr by storing it in an environment variable called
\code{BPSR_KEY}.
}
\examples{
\dontrun{
bps_set_key()
}

}
