% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subject.R
\name{bps_subject_category}
\alias{bps_subject_category}
\alias{bps_subject}
\title{Look up subject category and subject of dataset}
\usage{
bps_subject_category(domain_id = "0000", page = NULL, lang = c("ind", "eng"))

bps_subject(
  subject_category_id = NULL,
  domain_id = "0000",
  page = NULL,
  lang = c("ind", "eng")
)
}
\arguments{
\item{domain_id}{The domain ID of a BPS bureau. Defaults to "0000", which
is the domain ID of the central bureau. Use \code{\link[=bps_domain]{bps_domain()}} to see the
list of domain IDs.}

\item{page}{The page number of the result. Defaults to page one. Set to \code{Inf}
to request all pages.}

\item{lang}{The result's language. Must be either "ind" for Indonesian or
"eng" for English. Defaults to Indonesian.}

\item{subject_category_id}{The subject category ID. Use
\code{\link[=bps_subject_category]{bps_subject_category()}} to see the list of subject category IDs.}
}
\value{
\itemize{
\item \code{bps_subject_category()} returns a tibble containing subject category IDs
and titles.
\item \code{bps_subject()} returns a tibble containing subject IDs and titles, as well
as subject categories.
}
}
\description{
Get tables of dataset subject category and subject. The subject table
contains subject IDs, which are useful to request specific records in the
dataset table using \code{\link[=bps_dataset]{bps_dataset()}}.
\itemize{
\item \code{bps_subject_category()} requests the subject category table
\item \code{bps_subject()} requests the subject table
}
}
\examples{
\dontshow{if (httr2::secret_has_key("BPSR_KEY")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Look up the subject category
bps_subject_category(lang = "eng")

# Look up the subjects in the economic and trade category
bps_subject(subject_category_id = "2", lang = "eng")
\dontshow{\}) # examplesIf}
}
