% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-variables.R
\name{bps_vertical_var}
\alias{bps_vertical_var}
\alias{bps_derived_var}
\alias{bps_year}
\alias{bps_period}
\title{Look up dataset variable}
\usage{
bps_vertical_var(
  dataset_id = NULL,
  domain_id = "0000",
  page = NULL,
  lang = c("ind", "eng")
)

bps_derived_var(
  dataset_id = NULL,
  derived_var_group_id = NULL,
  domain_id = "0000",
  page = NULL,
  lang = c("ind", "eng")
)

bps_year(
  dataset_id = NULL,
  domain_id = "0000",
  page = NULL,
  lang = c("ind", "eng")
)

bps_period(
  dataset_id = NULL,
  domain_id = "0000",
  page = NULL,
  lang = c("ind", "eng")
)
}
\arguments{
\item{dataset_id}{The dataset ID. Use \code{\link[=bps_dataset]{bps_dataset()}} to see the list of
dataset IDs.}

\item{domain_id}{The domain ID of a BPS bureau. Defaults to "0000", which
is the domain ID of the central bureau. Use \code{\link[=bps_domain]{bps_domain()}} to see the
list of domain IDs.}

\item{page}{The page number of the result. Defaults to page one. Set to \code{Inf}
to request all pages.}

\item{lang}{The result's language. Must be either "ind" for Indonesian or
"eng" for English. Defaults to Indonesian.}

\item{derived_var_group_id}{The derived variable group ID. Use
\code{\link[=bps_derived_var]{bps_derived_var()}} to see the list of derived variable group IDs.}
}
\value{
\itemize{
\item \code{bps_vertical_var()} returns a tibble containing vertical variable IDs,
names and groups.
\item \code{bps_derived_var()} returns a tibble containing derived variable IDs,
names and groups.
\item \code{bps_year()} returns a tibble containing year IDs and years.
\item \code{bps_period()} returns a tibble containing period IDs and periods.
}
}
\description{
Get tables of dataset variables. Each table contains the variable IDs, which
are useful to request specific observations in a dataset using
\code{\link[=bps_get_dataset]{bps_get_dataset()}}.
\itemize{
\item \code{bps_vertical_var()} requests the vertical variable table
\item \code{bps_derived_var()} requests the derived variable table
\item \code{bps_year()} requests the year table
\item \code{bps_period()} requests the period table
}
}
\examples{
\dontshow{if (httr2::secret_has_key("BPSR_KEY")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Look up the vertical variables in the gross domestic product dataset
bps_vertical_var(dataset_id = "1955", lang = "eng")
\dontshow{\}) # examplesIf}
}
