% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{bps_view}
\alias{bps_view}
\title{Perform request to view endpoint}
\usage{
bps_view(model, id, ..., lang = c("ind", "eng"))
}
\arguments{
\item{model}{The result model, e.g. "publication" to view the details of
a particular publication. See the Model section below for more details.}

\item{id}{An ID, e.g. a publication ID.}

\item{...}{Name-value pairs that provide query parameters. See
\code{\link[httr2:req_url]{httr2::req_url_query()}} for more details.}

\item{lang}{The result's language. Must be either "ind" for Indonesian or
"eng" for English. Defaults to Indonesian.}
}
\value{
A list with the \code{bpsr_file} class.
}
\description{
The low-level function to send a request to the view endpoint.
}
\section{Model}{
\code{model} must be one of the following:
\itemize{
\item "glosarium";
\item "news";
\item "pressrelease";
\item "publication";
\item "statictable";
\item "tablestatistic";
\item "kbli2009";
\item "kbli2015";
\item "kbli2017";
\item "kbli2020"; or
\item "kbki2015".
}
}

\examples{
\dontshow{if (httr2::secret_has_key("BPSR_KEY")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Get information and the download URL of a spreadsheet dataset on annual
# inflation
bps_view("statictable", "915", domain = "0000", lang = "eng")
\dontshow{\}) # examplesIf}
}
\seealso{
\itemize{
\item \code{\link[=bps_download]{bps_download()}} to download resources by URL.
\item \code{\link[=bps_download_spreadsheet]{bps_download_spreadsheet()}} to download a spreadsheet dataset by ID.
}
}
