% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schYearFunctions.R
\name{bpsMonthDateCount}
\alias{bpsMonthDateCount}
\title{return a dataframe containing the number of school days in each month in BPS}
\usage{
bpsMonthDateCount(startYear = NA, endYear = NA)
}
\arguments{
\item{startYear}{a numeric value representing the first fiscal year to pull.
Note that a fiscal year is the second year in a school year, so the fiscal
year for 2022-23 would be 2023.}

\item{endYear}{a numeric value representing the last fiscal year to pull.
Note that a fiscal year is the second year in a school year, so the fiscal
year for 2022-23 would be 2023.}
}
\value{
a dataframe containing fiscal year, month, and the total school days
  in the month
}
\description{
This command returns a data frame containing the fiscal year, month, and
total school days each month in BPS
}
