% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schYearFunctions.R
\name{bpsSchedDaysOff}
\alias{bpsSchedDaysOff}
\title{Import the days off calendar from Github}
\usage{
bpsSchedDaysOff()
}
\value{
the days off data set, which includes calendar days off and an
  explanation for the event
}
\description{
Having quick access to a school year calendar is extremely valuable. To do
that, We need to have access to the days off so that we can remove them from
the calendar.
}
