% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schYearFunctions.R
\name{bpsScheduleDates}
\alias{bpsScheduleDates}
\title{Import a data set containing all scheduled school days in Boston Public
Schools for the upcoming school year}
\usage{
bpsScheduleDates(startYear = NA, endYear = NA)
}
\arguments{
\item{startYear}{a numeric value representing the first fiscal year to pull.
Note that a fiscal year is the second year in a school year, so the fiscal
year for 2022-23 would be 2023.}

\item{endYear}{a numeric value representing the last fiscal year to pull.
Note that a fiscal year is the second year in a school year, so the fiscal
year for 2022-23 would be 2023.}
}
\value{
a dataframe containing dates for the selected date range, including
  fiscYear, date, and day of the week.
}
\description{
This command creates a dataframe containing a list of all scheduled school
dates. It provides an option for a selective date range.
}
