% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnsOrgChartFunctions.R
\name{cashStringCleaner}
\alias{cashStringCleaner}
\title{Clean up strings with $ and , values embedded in them and convert them to
numeric values.}
\usage{
cashStringCleaner(vector)
}
\arguments{
\item{vector}{a string vector containing non-numeric character vectors
intended as cash values}
}
\value{
a numeric vector
}
\description{
Clean up strings with $ and , values embedded in them and convert them to
numeric values.
}
