% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titanFunctions.R
\name{convertTitanHTML}
\alias{convertTitanHTML}
\title{Convert an HTML report to a CSV}
\usage{
convertTitanHTML(inFile)
}
\arguments{
\item{inFile}{Path to the input file}
}
\value{
a csv file in the same directory as the html file
}
\description{
This function imports a Titan HTML report and converts it to a CSV
}
