% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titanFunctions.R
\name{convertWHTransfers}
\alias{convertWHTransfers}
\title{Clean and convert a Warehouse Transfers report to csv}
\usage{
convertWHTransfers(inFile)
}
\arguments{
\item{infile}{Path to the input file}
}
\value{
a csv file in the same directory as the html file
}
\description{
This function imports an HTML Warehouse Transfers Report, converts it to a
friendlier format, then exports it as a csv in the same directory.
}
