% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schYearFunctions.R
\name{fiscYearFromDate}
\alias{fiscYearFromDate}
\title{Extract the current fiscal year from a date value}
\usage{
fiscYearFromDate(date)
}
\arguments{
\item{date}{This is a vector containing date information}
}
\value{
the current school fiscal year for that date in numeric format
}
\description{
Since School years don't perfectly match to calendar years, this function is
designed to calculate the current school fiscal year from a date. The fiscal
year is a numeric value, meaning it's more useful than the SYXXXX school year
format.
}
