% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnsOrgChartFunctions.R
\name{fnsAvgLaborCost}
\alias{fnsAvgLaborCost}
\title{Imports the average labor hours by category for our staff}
\usage{
fnsAvgLaborCost()
}
\value{
a dataset containing average labor hours per year & the % multiplier
  for calculating benefits
}
\description{
Imports the average labor hours by category for our staff
}
