% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnsOrgChartFunctions.R
\name{fnsCoordList}
\alias{fnsCoordList}
\title{Import the full coordinator data set from Github}
\usage{
fnsCoordList()
}
\value{
the full, historical record of all coordinators at schools
}
\description{
This function provides quick access to the Coordinator List, which is
valuable for many tasks in Food & Nutrition Services.
}
