% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnsOrgChartFunctions.R
\name{fnsCoordListByDate}
\alias{fnsCoordListByDate}
\title{Import the current coordinator data set from Github}
\usage{
fnsCoordListByDate(cutOff = Sys.Date())
}
\value{
the currently assigned coordinators at schools
}
\description{
This function provides quick access to the Coordinator List, which is
valuable for many tasks in Food & Nutrition Services.
}
