% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnsOrgChartFunctions.R
\name{fnsCoordListCurrent}
\alias{fnsCoordListCurrent}
\title{Import the current coordinator data set from Github}
\usage{
fnsCoordListCurrent()
}
\value{
the currently assigned coordinators at schools
}
\description{
This function provides quick access to the Coordinator List, which is
valuable for many tasks in Food & Nutrition Services.
}
