% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titanFunctions.R
\name{fnsMealDesc}
\alias{fnsMealDesc}
\title{Import the meal description table from GitHub}
\usage{
fnsMealDesc()
}
\value{
a mergeable table containing program information
}
\description{
This function calls up the meal description crosswalk that can add a meal and
program name to any file.
}
