% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandaFunctions.R
\name{kinderStrToNum}
\alias{kinderStrToNum}
\title{Convert a character grade vector to a numeric grade vector}
\usage{
kinderStrToNum(vector)
}
\arguments{
\item{vector}{a character grade vector ('K0','K1','K2','01','02')}
}
\value{
a numeric grade vector (-2,-1,0,1,2)
}
\description{
This function takes a character grade vector or column and converts it to a
number For example, 'K0' is converted to -2, 'K2' is converted to 0, '09' to
9, and '11' to 11
}
