% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aceEndicoFunctions.R
\name{loadEndicoReports}
\alias{loadEndicoReports}
\title{Imports Ace Endico Reports listed in a directory as well as the allowance
sheets.}
\usage{
loadEndicoReports(directory)
}
\arguments{
\item{directory}{A directory where the Ace Endico Reports live.}
}
\value{
Two data sets, one containing all items in the Ace Endico reports,
  one containing all of the allowances that are listed on sheets 2-5 of an
  Ace Endico Report.
}
\description{
This function pulls in a collection of processed Ace Endico reports
}
