% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/claimsFunctions.R
\name{readCACFPAttendance}
\alias{readCACFPAttendance}
\title{Import data stored in the CACFP attendance Google Sheets and simplify it down
to the CACFP ID & attendance value.}
\usage{
readCACFPAttendance(sheetID, dataSheet)
}
\arguments{
\item{sheetID}{The end of a google sheet url. Usually a 44'ish long
character string that uniquely identifies a sheet.}

\item{dataSheet}{The sheet in the spreadsheet that contains the attendance
data.}
}
\value{
A cleaned up data frame with 3 columns - the school's cacfp id, date,
  and attendance
}
\description{
This function imports and lightly cleans attendance data that we currently
store & manage inside of a Google sheet in a really wonky format. This piece
of scripting could very easily fail if we fail to maintain rigorous standards
around how we format those tables.
}
