% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/claimsFunctions.R
\name{readCACFPAttendanceCount}
\alias{readCACFPAttendanceCount}
\title{Read the school-level attendance sheets that schools use for CACFP}
\usage{
readCACFPAttendanceCount(sheetID, combinePrograms = T)
}
\arguments{
\item{sheetID}{The end of a google sheet url. Usually a 44'ish long character
string that uniquely identifies a sheet.}

\item{combinePrograms}{A boolean that indicates whether or not attendance
should be combined accross sheets. Some programs use multiple sheets for 1
program, in cases like these you'd want combine programs to be True. In
cases where multiple sheets are used for multiple, distinct programs, this
should be set to false to count each independently.}
}
\value{
A cleaned up data frame with 3 columns - the school's cacfp id, date,
  and attendance
}
\description{
This function imports data from a Google sheet that tracks attendance at
schools, tallies the total attendance by program by day
}
