% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titanFunctions.R
\name{removeDiscontinued}
\alias{removeDiscontinued}
\title{Remove DISCONTINUED string from the Identifier column}
\usage{
removeDiscontinued(vector)
}
\arguments{
\item{vector}{A vector to remove the "Discontinued" string from}
}
\value{
A vector with 'Discontinued' removed from all of the strings. More
  specifically, a vector with everything after the first space removed.
}
\description{
This function addresses the bad data practice of altering product keys in our
inventory database by writing DISCONTINUED after them when they cease using
them. It's pretty simple, and removes everything after the first space in a
string.
}
