% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schYearFunctions.R
\name{schToFiscalYear}
\alias{schToFiscalYear}
\title{Convert a string school year value into a numeric fiscal year value}
\usage{
schToFiscalYear(schYear)
}
\arguments{
\item{schYear}{a string school year value formatted "SY0000"}
}
\value{
A vector containing the numeric Fiscal year value, which is equal to
  the second part of the school year + 2000
}
\description{
Since the string format for dates is frankly bad for working with time, this
turns the school year value "SY0000" format into a numeric value equal to the
second digits + 2000. Which means that if, somehow, this package is still
being deployed in 2100, you'll need to make this more dynamic somehow.
}
