% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schYearFunctions.R
\name{schYearFromDate}
\alias{schYearFromDate}
\title{Extract the current school year from a date value}
\usage{
schYearFromDate(date)
}
\arguments{
\item{date}{This is a vector containing date information}
}
\value{
the current school year for that date in string format
}
\description{
Since School years don't perfectly match to calendar years, this function is
designed to calculate the current school year from a date. School Year is a
string value, formatted 'SY0000', where the first set of zeroes indicate the
year when a school year starts, and the second set indicates the year a
school year ends.
}
