% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titanFunctions.R
\name{sepMultipleInventory}
\alias{sepMultipleInventory}
\title{Separate and create multiple line items in the Purchasing report in instances
where a report combines line items in this convention - 'x unit & y case'}
\usage{
sepMultipleInventory(df, var)
}
\arguments{
\item{df}{A Titan report containing lines with units and cases listed
together in one line item.}

\item{var}{The variable in which cases and units are listed together
formatted 'x units & y cases'.}
}
\value{
a new dataset where the units and cases are split into multiple line
  items.
}
\description{
Titan designs some of its reports, most notably the purchasing report, to
list units and cases in the same line item. In order to get around this
convention, this command will split those items and produce separate line
items, one for ases, and one for units.
}
