% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titanFunctions.R
\name{splitNameAndID}
\alias{splitNameAndID}
\title{Convert a Titan building name into a school name and its constituent ID}
\usage{
splitNameAndID(df, var)
}
\arguments{
\item{df}{A Titan report containing a building variable}

\item{var}{The variable containing the Titan school names}
}
\value{
a data frame with the school ID separated from the school name
}
\description{
We updated titan building names to contain IDs behind a # mark. This code
splits the Titan IDs into a building name and a school code
}
