% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schYearFunctions.R
\name{tableExpanderByDay}
\alias{tableExpanderByDay}
\title{Take one of Dane's collapsed tables by day that needs to be expanded to merge,
and expands it}
\usage{
tableExpanderByDay(condensedDF, startDateVar, endDateVar)
}
\arguments{
\item{condensedDF}{This is a data frame like the one described above, where a
status is assigned to an entity by day, but the table is maintained such
that there is an initial day and final day column. Excludes non-school days!}

\item{startYearVar}{The variable containing the initial day of the status}

\item{endYearVar}{The variable containin the final day of the status}
}
\value{
a longform table with each interstitial day that can be merged into
  an existing report
}
\description{
When Dane maintains tables, he likes to keep them fairly simple to maintain.
For example, when he's tracking something that might change per annum, he
likes to build tables with two year columns, one for the initial year that
the assigned status occurs and one year for the final year of the status. If
the second year is empty, it means that the status is still current. This
algorithm takes these condensed tables and converts them into long form
tables that can be merged on a fiscal year variable.
}
