% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titanFunctions.R
\name{titanConvertCashRec}
\alias{titanConvertCashRec}
\title{Convert a daily cash reconcilation report to a csv in the same directory}
\usage{
titanConvertCashRec(inFile)
}
\arguments{
\item{inFile}{the path to the input file}
}
\value{
creates a csv file in the same directory as the html file
}
\description{
This function imports an HTML cash reconciliation report and converts it to a
csv
}
