% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titanFunctions.R
\name{titanConvertEditCheck}
\alias{titanConvertEditCheck}
\title{Convert an HTML Edit Check Report to a CSV in the same directory}
\usage{
titanConvertEditCheck(inFile)
}
\arguments{
\item{inFile}{Path to the input file}
}
\value{
a csv file in the same directory as the html file
}
\description{
This function imports a Titan HTML Edit Check report and converts it to a csv
}
