% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titanFunctions.R
\name{titanConvertProdRecs}
\alias{titanConvertProdRecs}
\title{Convert a production record file to a csv}
\usage{
titanConvertProdRecs(inFile, htmlListVars = c("school", "date"))
}
\arguments{
\item{inFile}{Path to the html production records report}
}
\value{
a csv file in the same directory as the html file
}
\description{
This function imports a Titan Production Records report & converts it to a
CSV
}
