% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titanFunctions.R
\name{titanConvertReceiptTix}
\alias{titanConvertReceiptTix}
\title{Clean and convert a Receipt Tickets report to csv}
\usage{
titanConvertReceiptTix(inFile)
}
\arguments{
\item{inFile}{Path to the input file}
}
\value{
a csv file in the same directory as the html file
}
\description{
This function imports an HTML Receipt Tickets Report, converts it to a
friendlier format, then exports it as a csv in the same directory.
}
