% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titanFunctions.R
\name{titanEditCheckAFCleaner}
\alias{titanEditCheckAFCleaner}
\title{Make specific adjustments to an Edit Check File's attendance and enrollment
and adjusts some school data.}
\usage{
titanEditCheckAFCleaner(ec)
}
\arguments{
\item{ec}{an edit check report being imported to R}
}
\value{
an edit check report with attendance values properly adjusted and
  school designations appropriately aligned
}
\description{
This function adjusts enrollment and attendance on an Edit Check report. Some
schools also need to have their meals assigned to other schools, and this
does that as well. It uses a couple of tables in my GitHub to accomplish
this, then adjusts some values throughout the EC report based on the values
of those tables.
}
