% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titanFunctions.R
\name{titanImportCashRec}
\alias{titanImportCashRec}
\title{Convert an HTML daily cash reconciliation report from Titan into an R Data
frame}
\usage{
titanImportCashRec(inFile)
}
\arguments{
\item{inFile}{Path to the input file}
}
\value{
an R dataframe containing the data from the HTML cash reconciliation
  report
}
\description{
This function imports an HTML daily cash reconciliation report in HTML and
converts it to a format for use in R
}
