% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titanFunctions.R
\name{titanImportEditCheck}
\alias{titanImportEditCheck}
\title{Import a Titan Edit Check Report}
\usage{
titanImportEditCheck(inFile)
}
\arguments{
\item{inFile}{Path to the Edit Check file}
}
\value{
a cleaned edit check report dataframe
}
\description{
This function imports and transforms an HTML Edit Check report so that users
can interact with it like a normal data set.
}
