% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titanFunctions.R
\name{titanImportHTML}
\alias{titanImportHTML}
\title{Import a Titan HTML report}
\usage{
titanImportHTML(inFile)
}
\arguments{
\item{inFile}{Path to the saved Titan HTML report}
}
\value{
A data frame containing the content of the Titan Report
}
\description{
This function simplifies the process of importing and cleaning a report from
Titan
}
