% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titanFunctions.R
\name{titanImportProdRecs}
\alias{titanImportProdRecs}
\title{Import a Titan HTML Production Report}
\usage{
titanImportProdRecs(inFile, htmlListVars = c("school", "date"))
}
\arguments{
\item{inFile}{The path to a saved Titan Production Records Report}

\item{htmlListVars}{A vector containing the variables that you grouped the
production records by. must be variables like "school" and "date"}
}
\value{
a data frame containing the content of a Production records report
}
\description{
This function imports and cleans an HTML production records report from Titan
}
