% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titanFunctions.R
\name{titanImportReceiptTix}
\alias{titanImportReceiptTix}
\title{Import a Titan HTML Item Receipts}
\usage{
titanImportReceiptTix(inFile)
}
\arguments{
\item{inFile}{The path to a saved Titan Receipt Tickets Report}
}
\value{
a data frame containing the content of a Receipt Tickets Report
}
\description{
This function imports and cleans an HTML production records report from Titan
}
