% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titanFunctions.R
\name{titanImportWHTransfers}
\alias{titanImportWHTransfers}
\title{Import and clean the Warehouse transfers report}
\usage{
titanImportWHTransfers(inFile)
}
\arguments{
\item{inFile}{This is the directory of an HTML Warehouse Transfers Report}
}
\value{
A data frame with the product and identifier values appropriately placed
}
\description{
This function imports the badly formatted Titan Warehouse Transfer Report and
moves the product and product identifier to the same row as the transfer
reports.
}
