% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schYearFunctions.R
\name{yyyymmddString}
\alias{yyyymmddString}
\title{Build a date string formated YYYYMMDD}
\usage{
yyyymmddString(date)
}
\arguments{
\item{date}{the day you want to convert to the string}
}
\value{
a string with the current date formatted YYYYMMDD
}
\description{
This function returns a date string formatted YYYYMMDD for use in file
directories & file names.
}
