% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schYearFunctions.R
\name{yyyymmString}
\alias{yyyymmString}
\title{Build a current date string formated YYYY-MM}
\usage{
yyyymmString(date)
}
\arguments{
\item{date}{the day you want to convert to the string}
}
\value{
a string with the date formatted YYYY-MM
}
\description{
This function returns a date string formatted YYYY-MM for use in file
directories & file names.
}
